fix: ensure proper checking of "in-use" status when fetching metadata
This PR fixes a bug where users could have orphaned metalmachines that still
had ownerrefs to a server that had alredy been made available for another
cluster. This led to incorrect metadata being returned. We now bail and
error early if we encounter multiple ownerrefs in the list pointing to
the same server.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>