fix: handle conflicts on server update
The problem here was that there were two `Update()`s, one for `Status`
and another one for the `Spec` essentially. `Status` update could have
succeeded and marked server as `InUse`, while `Spec` update failed on
conflict error.

As we ignore conflict errors one level up, server was marked as `InUse`
while never attached to any `MetalMachine`, so it was left as orphaned.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>