feat: power down servers !InUse && IsClean
Basically this repeats the power off action which should be done by the
agent image. In case of Sfyra and QEMU VMs, they can't be shutdown via
ACPI, so this code is going to bring Sfyra case closer to the bare metal
case, plus it should handle power off if Server is registered manually
(without agent).

Also mark servers which are `InUse && !IsClean` as `Ready` to assert
that all the servers are ready in the test.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>