feat: support server acceptance
This PR adds a bool field to the server spec of `accepted`. This allows
users to adopt a server that has been registered into the sidero
environment. Doing so allows for a server's full lifecycle to be
managed. Users can either edit a server resource to update the accepted
field, or create the server resource manually and supply that with their
spec.

The PR also adds some edge case handling around the sidero API server to
check to make sure a server is accepted before allowing a wipe directive
to be returned, as well as a check for acceptance when trying to use a
serverref in the sidero capi provider.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>