feat: don't power off the server in discovery agent
This disables machine power off in discovery agent.

If `Server` resource has `IPMI` information, it will be powered off by
the Sidero controller once it's clean and accepted.

This simplifies the flow for non-IPMI environments, as machines are now
sitting in the boot loop and might start provisioning as soon as
`Server` is allocated.

Fixes #204

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>