feat: inject iPXE script into the iPXE binaries
Fixes #227

See also https://github.com/talos-systems/pkgs/pull/267

The rough idea is that iPXE binaries we're building contain placeholder
script which is replaced on the fly with the actual script which depends
on Sidero API endpoint. Patching is a bit tricky for `undionly.kpxe`, as
it is compressed, so we patch uncompressed version and compress the
final asset after that.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>