docs: handle latest version banner and canonical urls
This PR is a copy of two PRs from talos:

https://github.com/siderolabs/talos/pull/5389/files
https://github.com/siderolabs/talos/pull/5391/files

It ensures that the banner is shown if the user isn't viewing the latest
sidero docs. This actually already was working but the banner html was in an incorrect spot, so moved that to match what's seen in Talos repo. Also adds the canonical links for search engines to update their
indexes.

Signed-off-by: Spencer Smith <spencer.smith@talos-systems.com>