fix: multi doc machine config generation
The very last step in the machine config generation is to apply a
specific node label to match the Kubernetes node back to the Machine.

This step was using legacy JSON patches which rendered impossible to use
multi-doc machine config.

Rebuild to use strategic patch instead, this changes formatting of the
config in the tests, but not in real production.

Also this means that Sidero is locked to the Talos version it was built
for (it won't support newer Talos machine configs), but that has already
been the case with strategic patches.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>